Split models and added the ability to load a TurvableItem from a commaSeperatedString
